requirements.txt
openai
streamlit
loguru
pandas
python-dotenv
